refactor: layout redesign, drawer system and page standardization
